requirements.txt
google-genai
tweepy
praw
feedparser
Pillow
pytrends
tenacity
